[Android] Fix for browser tests
Chromium change:
https://chromium.googlesource.com/chromium/src/+/be2eae1c83be911e7154acef8a86660482918cdf

[Shortcuts Widget] Implemented the UI of the small version of the widget

This CL introduces a second version of the Quick Action Search Widget that is one row tall and doesn't include dino.

With this change, two versions of the Quick Action Search Widget will appear in the widget picker (medium and small).  Note that this CL doesn't include dynamically resizing the widget to switch between medium and small layouts, that will come in a later CL.

Screenshot: http://screen/C6NuGTaFfZXkcER.png

Bug: 1213541