[Android] Fixes for unit tests
Chromium change:
https://chromium.googlesource.com/chromium/src/+/9bb17bfbf232dcc3365a3b781ef62ae18e747d23

Add Android setting for HTTPS-First Mode

This adds a new setting to the "Privacy and Security" settings subpage
for HTTPS-First Mode, and hooks it up to the native pref and enterprise
policy.

Note that new code added in this CL refers to this as "HTTPS-First Mode"
but still references or calls into old code that still refers to this as
"HTTPS-Only Mode". A followup CL will handle more of the renaming.

Bug: 1218526